test(browser-tools): move sharp metadata-path guard to CI-run .mjs suite
